requirements.txt
PyQt6>=6.5.0
PyQt6-Qt6>=6.8.0
pillow>=9.5.0
pywin32>=306; sys_platform == 'win32'
pyobjc-framework-Cocoa>=9.0; sys_platform == 'darwin'
